finance_assistant/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-bcrypt==1.0.1
pandas==2.1.4
numpy==1.26.0
scikit-learn==1.3.2
nltk==3.8.1
matplotlib==3.8.2
seaborn==0.13.0
reportlab==4.0.6
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.0
pyjwt==2.8.0
werkzeug==3.0.1
jinja2==3.1.2